NEWRELIC-7888: (nestjs) add a Nest.js example app
This is a basic Nest.js app using Prisma for the database and Winston
for logging. It exercises several features of Nest.js like
interceptors and guards. Like other examples here, it can be run in
Docker.

Optionally, via  NES_USE_FASTIFY envvar, it may use Fastify, but it
defaults to Express, as that's Nest.js's default.